Remove standalone Javascript includes
Asset precompilation was breaking in production because of standalone
javascripts (those included with the `javascript_includes` block and not
listed in the application.js manifest) were skipping the precompilation
stage. We don't want to enable live compilation in prod. for a host of
reasons, and in the end reducing HTTP requests by minifying into a
single file is a big win.

It is still true that certain scripts are meant to run only a single
page. However, they can safely be included in the application manifest
because our javascript pattern tends to define a number of helper
functions, and bind to a specific element/DOM event at run time (ex
`.reply-btn` click), causing a no-op on other pages.

We will need to come up with an alternate solution if we run into a use
case where we truly cannot include a script in the manifest.